Context awareness merging finished

+++ src/main/java/de/bnder/taskmanager/commands/task/DeleteTask.java
@@ -24,42 +24,46 @@ public class DeleteTask {
     public static void deleteTask(Member member, TextChannel textChannel, String taskID, SlashCommandEvent slashCommandEvent) {
         final Locale langCode = Localizations.getGuildLanguage(member.getGuild());
         final String embedTitle = Localizations.getString("task_message_title", langCode);
-        if (PermissionSystem.hasPermission(member, TaskPermission.DELETE_TASK)) {
-            if (!taskID.equalsIgnoreCase("done")) {
-                final de.bnder.taskmanager.utils.Task task = new de.bnder.taskmanager.utils.Task(taskID, textChannel.getGuild());
-                if (task.exists()) {
-                    task.delete();
-                    MessageSender.send(embedTitle, Localizations.getString("task_deleted", langCode, new ArrayList<>() {
-                        {
-                            add(taskID);
-                        }
-                    }), textChannel, Color.green, langCode, slashCommandEvent);
+        if (taskID != null) {
+            if (PermissionSystem.hasPermission(member, TaskPermission.DELETE_TASK)) {
+                if (!taskID.equalsIgnoreCase("done")) {
+                    final de.bnder.taskmanager.utils.Task task = new de.bnder.taskmanager.utils.Task(taskID, textChannel.getGuild());
+                    if (task.exists()) {
+                        task.delete();
+                        MessageSender.send(embedTitle, Localizations.getString("task_deleted", langCode, new ArrayList<>() {
+                            {
+                                add(taskID);
+                            }
+                        }), textChannel, Color.green, langCode, slashCommandEvent);
+                    } else {
+                        MessageSender.send(embedTitle, Localizations.getString("no_task_by_id", langCode, new ArrayList<>() {
+                            {
+                                add(taskID);
+                            }
+                        }), textChannel, Color.red, langCode, slashCommandEvent);
+                    }
                 } else {
-                    MessageSender.send(embedTitle, Localizations.getString("no_task_by_id", langCode, new ArrayList<>() {
-                        {
-                            add(taskID);
+                    try {
+                        for (QueryDocumentSnapshot boardDoc : Main.firestore.collection("server").document(member.getGuild().getId()).collection("boards").get().get()) {
+                            for (QueryDocumentSnapshot taskDoc : boardDoc.getReference().collection("user-tasks").whereEqualTo("status", 2).get().get()) {
+                                taskDoc.getReference().delete();
+                            }
                         }
-                    }), textChannel, Color.red, langCode, slashCommandEvent);
+                        for (QueryDocumentSnapshot groupDoc : Main.firestore.collection("server").document(member.getGuild().getId()).collection("groups").get().get()) {
+                            for (QueryDocumentSnapshot taskDoc : groupDoc.getReference().collection("group-tasks").whereEqualTo("status", 2).get().get()) {
+                                taskDoc.getReference().delete();
+                            }
+                        }
+                    } catch (InterruptedException | ExecutionException e) {
+                        logger.error(e);
+                    }
+                    MessageSender.send(embedTitle, Localizations.getString("deleted_done_tasks", langCode), textChannel, Color.green, langCode, slashCommandEvent);
                 }
             } else {
-                try {
-                    for (QueryDocumentSnapshot boardDoc : Main.firestore.collection("server").document(member.getGuild().getId()).collection("boards").get().get()) {
-                        for (QueryDocumentSnapshot taskDoc : boardDoc.getReference().collection("user-tasks").whereEqualTo("status", 2).get().get()) {
-                            taskDoc.getReference().delete();
-                        }
-                    }
-                    for (QueryDocumentSnapshot groupDoc : Main.firestore.collection("server").document(member.getGuild().getId()).collection("groups").get().get()) {
-                        for (QueryDocumentSnapshot taskDoc : groupDoc.getReference().collection("group-tasks").whereEqualTo("status", 2).get().get()) {
-                            taskDoc.getReference().delete();
-                        }
-                    }
-                } catch (InterruptedException | ExecutionException e) {
-                    logger.error(e);
-                }
-                MessageSender.send(embedTitle, Localizations.getString("deleted_done_tasks", langCode), textChannel, Color.green, langCode, slashCommandEvent);
+                MessageSender.send(embedTitle, Localizations.getString("need_to_be_serveradmin_or_have_admin_permissions", langCode), textChannel, Color.red, langCode, slashCommandEvent);
             }
         } else {
-            MessageSender.send(embedTitle, Localizations.getString("need_to_be_serveradmin_or_have_admin_permissions", langCode), textChannel, Color.red, langCode, slashCommandEvent);
+            MessageSender.send(embedTitle, Localizations.getString("context_awareness_no_task_id_found", langCode), textChannel, Color.red, langCode, slashCommandEvent);
         }
     }
